updated tests
stub for documentation

did merge and added testing support for structure in the harness

made reference to new structure test harness property
