+++ src/Machete.HL7.Tests/Machete.HL7.Tests.csproj
@@ -49,7 +49,19 @@
     <Compile Include="DateTimeExtensionsTests.cs" />
     <Compile Include="Query_Specs.cs" />
     <Compile Include="SchemaParser_Specs.cs" />
+    <Compile Include="Segments\DateTimeSegment.cs" />
+    <Compile Include="Segments\DateTimeSegmentMap.cs" />
+    <Compile Include="Segments\EVNSegment.cs" />
+    <Compile Include="Segments\EVNSegmentMap.cs" />
+    <Compile Include="Segments\MessageLayout.cs" />
+    <Compile Include="Segments\MessageLayoutMap.cs" />
+    <Compile Include="Segments\MSG.cs" />
+    <Compile Include="Segments\MSGComponentMap.cs" />
     <Compile Include="Segments\MSHSegment.cs" />
+    <Compile Include="Segments\MSHSegmentMap.cs" />
+    <Compile Include="Segments\OptionalMessageLayout.cs" />
+    <Compile Include="Segments\OptionalMessageLayoutMap.cs" />
+    <Compile Include="Segments\TestHL7Entity.cs" />
     <Compile Include="Template_Specs.cs" />
     <Compile Include="Tests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
