+++ src/Machete.HL7.Tests/Query_Specs.cs
@@ -7,12 +7,12 @@
 
     [TestFixture]
     public class Querying_a_schema :
-        HL7MacheteTestHarness<MSHSegment, HL7Entity>
+        HL7MacheteTestHarness<TestHL7Entity, HL7Entity>
     {
         [Test]
         public void Should_be_possible()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234|||K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234|||K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
 
@@ -28,7 +28,7 @@
         [Test]
         public void Should_not_blow_up()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234|||K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234|||K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
 
@@ -44,7 +44,7 @@
         [Test]
         public void Should_throw_ValueMissingException()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234|||K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234|||K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
 
@@ -64,7 +64,7 @@
         [Test]
         public void Should_parse_a_series_of_segments()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||ORU^R01|K113|P|
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234||ORU^R01|K113|P|
 EVN|A08|201701131234|||12901";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
@@ -87,7 +87,7 @@ EVN|A08|201701131234|||12901";
         [Test]
         public void Should_be_able_to_skip_segments_via_selection()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||ORU^R01|K113|P|
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234||ORU^R01|K113|P|
 EVN|A08|201701131234|||12901";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
@@ -109,7 +109,7 @@ EVN|A08|201701131234|||12901";
         [Test]
         public void Should_parse_a_series_of_segments_using_optional()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||ORU^R01|K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234||ORU^R01|K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
 
@@ -131,7 +131,7 @@ EVN|A08|201701131234|||12901";
         [Test]
         public void Should_parse_a_series_of_segments_but_not_match()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||ORU^R01|K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234||ORU^R01|K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
 
@@ -146,7 +146,7 @@ EVN|A08|201701131234|||12901";
         [Test]
         public void Should_parse_a_segment_and_parse_into_the_component()
         {
-            const string message = @"MSH|^~\&|LIFTLAB||UBERMED||201701131234||ORU^R01|K113|P|";
+            const string message = @"MSH|^~\&|MACHETELAB||UBERMED||201701131234||ORU^R01|K113|P|";
 
             Parsed<HL7Entity> parsed = Parser.Parse(message);
