+++ src/Machete.HL7.Tests/Segments/DateTimeSegment.cs
@@ -1,0 +1,11 @@
+﻿namespace Machete.HL7.Tests.Segments
+{
+    using System;
+
+    public interface DateTimeSegment :
+        HL7Segment
+    {
+        Value<DateTimeOffset> TestDateTimeOffset { get; }
+        Value<DateTimeOffset> TestDateTimeOffsetWithTime { get; }
+    }
+}
