+++ src/Machete.HL7.Tests/Segments/EVNSegmentMap.cs
@@ -1,0 +1,25 @@
+﻿namespace Machete.HL7.Tests.Segments
+{
+    public class EVNSegmentMap :
+        HL7SegmentMap<EVNSegment, HL7Segment>
+    {
+        public EVNSegmentMap()
+        {
+            Id = "EVN";
+
+            Name = "Event Type";
+
+            Value(x => x.EventTypeCode, 1, x => x.MaxLength = 3);
+            Value(x => x.RecordedDateTime, 2, x => x.LongDateTime().Required());
+            Value(x => x.DateTimeOfPlannedEvent, 3);
+            Value(x => x.EventReasonCode, 4, x =>
+            {
+                x.MaxLength = 3;
+                x.Required();
+            });
+//            Value(x => x.OperatorId, 5).MaxLength(250);
+            Value(x => x.EventOccurred, 6, x => x.LongDateTime());
+            //          Value(x => x.EventFacility, 7).MaxLength(241);
+        }
+    }
+}
