+++ src/Machete.HL7.Tests/Segments/MessageLayout.cs
@@ -1,0 +1,9 @@
+﻿namespace Machete.HL7.Tests.Segments
+{
+    public interface MessageLayout :
+        HL7Layout
+    {
+        Segment<MSHSegment> MSH { get; }
+        Layout<OptionalMessageLayout> Optional { get; }
+    }
+}
