+++ src/Machete.HL7.Tests/Segments/OptionalMessageLayout.cs
@@ -1,0 +1,8 @@
+﻿namespace Machete.HL7.Tests.Segments
+{
+    public interface OptionalMessageLayout :
+        HL7Layout
+    {
+        Segment<EVNSegment> EVN { get; }
+    }
+}
