+++ src/Machete.HL7.Tests/Segments/TestHL7Entity.cs
@@ -1,0 +1,8 @@
+namespace Machete.HL7.Tests.Segments
+{
+    public interface TestHL7Entity :
+        HL7Entity
+    {
+        
+    }
+}
