+++ src/Machete/Testing/IMacheteTestHarness.cs
@@ -5,5 +5,6 @@ namespace Machete.Testing
     {
         ISchema<TSchema> Schema { get; }
         IParser<TSchema> Parser { get; }
+        IStructure<TSchema> Structure { get; }
     }
 }
