+++ src/Machete/ValueExtensions.cs
@@ -1,7 +1,6 @@
 ﻿namespace Machete
 {
     using System;
-    using System.Diagnostics;
     using System.Threading.Tasks;
